docs:update behaviors docs
