+++ packages/graphin/docs/behaviros/canvas.en-US.md
@@ -1,0 +1,41 @@
+---
+title: Canvas interaction
+order: 1
+group:
+  path: /behaviors
+  title: Interactive behavior
+  order: 3
+nav:
+  path: /graphin
+  order: 1
+---
+
+## DragCanvas
+
+> Drag the canvas
+
+<API src='../../src/behaviors/DragCanvas.tsx'>
+
+## ZoomCanvas
+
+Zoom Canvas
+
+<API src='../../src/behaviors/ZoomCanvas.tsx'>
+
+## BrushSelect
+
+brush select the canvas
+
+<API src='../../src/behaviors/BrushSelect.tsx'>
+
+## LassoSelect
+
+lasso select canvas
+
+<API src='../../src/behaviors/LassoSelect.tsx'>
+
+## ResizeCanvas
+
+resize canvas
+
+<API src='../../src/behaviors/ResizeCanvas.tsx'>
