+++ packages/graphin/docs/behaviros/canvas.md
@@ -10,22 +10,47 @@ nav:
   order: 1
 ---
 
-# DragCanvas 拖拽画布
+## 用法
+
+```tsx | pure
+import React from 'react';
+import Graphin, { Utils, Behaviors, GraphinContext } from '@antv/graphin';
+
+const data = Utils.mock(10)
+  .random()
+  .graphin();
+
+const { DragCanvas, ZoomCanvas, DragNode, ActivateRelations } = Behaviors;
+
+export default () => {
+  return (
+    <div>
+      <Graphin data={data}>
+        <ZoomCanvas enableOptimize />
+        <DragNode />
+        <ActivateRelations trigger="click" />
+      </Graphin>
+    </div>
+  );
+};
+```
+
+## DragCanvas 拖拽画布
 
 <API src='../../src/behaviors/DragCanvas.tsx'>
 
-# ZoomCanvas 缩放画布
+## ZoomCanvas 缩放画布
 
 <API src='../../src/behaviors/ZoomCanvas.tsx'>
 
-# BrushSelect 圈选画布
+## BrushSelect 圈选画布
 
 <API src='../../src/behaviors/BrushSelect.tsx'>
 
-# LassoSelect 拉索画布
+## LassoSelect 拉索画布
 
 <API src='../../src/behaviors/LassoSelect.tsx'>
 
-# ResizeCanvas Resize 画布
+## ResizeCanvas Resize 画布
 
 <API src='../../src/behaviors/ResizeCanvas.tsx'>
